+++ src/app/home/FollowContent.tsx
@@ -1,9 +1,10 @@
 import { useSearchParams } from 'next/navigation';
 import { useFollowMembers } from '@/apis/follow';
 import { type FollowMemberType } from '@/apis/schema/member';
-import FollowMissionList, { FollowMissionListSkeleton } from '@/app/home/FollowMissionList';
-import FollowSummary, { FollowSummarySkeleton } from '@/app/home/FollowSummary';
-import MissionList from '@/app/home/MissionList';
+import FollowMissionList from '@/app/home/Content/FollowMissionList';
+import FollowSummary, { FollowSummarySkeleton } from '@/app/home/Content/FollowSummary';
+import HomeMissionList from '@/app/home/Content/MissionList';
+import MissionList from '@/components/MissionList';
 import { flex } from '@styled-system/patterns';
 
 function FollowContent() {
@@ -12,7 +13,7 @@ function FollowContent() {
   if (!isFollower)
     return (
       <div className={containerCss}>
-        <MissionList />
+        <HomeMissionList />
       </div>
     );
 
@@ -20,7 +21,11 @@ function FollowContent() {
     return (
       <div className={containerCss}>
         <FollowSummarySkeleton />
-        <FollowMissionListSkeleton />
+
+        <h2 className={headingCss}>
+          <span>미션 목록</span>
+        </h2>
+        <MissionList.Skeleton />
       </div>
     );
   }
@@ -28,12 +33,22 @@ function FollowContent() {
   return (
     <div className={containerCss}>
       <FollowSummary {...selectedFollowData} />
+
+      <h2 className={headingCss}>
+        <span>미션 목록</span>
+      </h2>
       <FollowMissionList followId={selectedFollowData.memberId} />
     </div>
   );
 }
 
 export default FollowContent;
+
+const headingCss = flex({
+  padding: '12px 4px',
+  textStyle: 'body4',
+  color: 'text.primary',
+});
 
 const useGetSelectFollowData = (): { data: FollowMemberType | null; isLoading: boolean; isFollower: boolean } => {
   const { data, isLoading } = useFollowMembers();
