+++ src/components/MissionList/index.tsx
@@ -2,12 +2,14 @@ import MissionEmptyList from './Empty';
 import MissionItem from './Item';
 import MissionLinkItem from './LinkItem';
 import MissionListContainer from './ListContainer';
+import MissionListNoticeEmpty from './NoticeEmpty';
 import MissionListSkeleton from './Skeleton';
 
 export default Object.assign(MissionListContainer, {
   Container: MissionListContainer,
   item: MissionItem,
   LinkItem: MissionLinkItem,
-  Empty: MissionEmptyList,
   Skeleton: MissionListSkeleton,
+  SuggestAdd: MissionEmptyList,
+  NoticeEmpty: MissionListNoticeEmpty,
 });
